requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-celery-results==2.5.1
couchdb==1.2
celery==5.3.6
redis==5.0.3
opencv-python-headless==4.9.0.80
numpy==2.1.0
pdf2image==1.17.0
reportlab==4.2.5
qrcode==7.4.2
Pillow==10.4.0
pyzbar==0.1.9
